Move value types into their own namespace as per code review comments.
